subido mensaje al eliminar Session o WeeklyRoutine

+++ backend/src/main/java/com/gym/backend/presenter/SessionPresenter.java
@@ -14,6 +14,7 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
 import com.gym.backend.Response;
+import com.gym.backend.business.services.RoutineDayService;
 import com.gym.backend.business.services.SessionService;
 import com.gym.backend.business.services.UserService;
 import com.gym.backend.model.Session;
@@ -25,6 +26,9 @@ public class SessionPresenter {
 
     @Autowired
     private SessionService sessionService;
+
+    @Autowired
+    private RoutineDayService routineDayService;
 
     @Autowired
     private UserService userService;
@@ -120,6 +124,11 @@ public class SessionPresenter {
 
         if (!session.getUser().getId().equals(user.getId()))
             return Response.dbError("No puede eliminar una sesión que no le pertenece.");
+
+        // Validar si la sesión está asociada a algún RoutineDay
+        if (routineDayService.existsBySessionId(id)) {
+            return Response.dbError("No se puede eliminar la sesión porque está asociada a una rutina diaria.");
+        }
 
         sessionService.delete(id);
         return Response.ok("La sesión con ID " + id + " fue eliminada correctamente.");
